Use `tokio::io::split()` for windows workaround
Because mio_serial has a bug on windows, where it errors when writing to a cloned serial port,
this workaround uses `tokio`s `io::split()` to get a reader and writer handle from a single serial port.
The writer is then used with a `FramedWriter` and unbounded channel to be able to send bytes to
the serial port asynchronously.

Updated the dependencies `crossterm`, which fixes a bug where some characters were not received
by the `EventStream` (ex. `@` on a keyboard layout where to type it the key-combination `AltGr+2`
is used).

Updated `tokio-serial` to latest version (fixes rust-analyzer crash where it couldn't parse the version).